Remove RemoteStore (#46)
- Removed `RemoteStore`.
- Functionality of `RemoteStore` is moved into `Store`.
- Changed `Store.cleanup()` to `Store.close()` to be more inline with
  sockets, files, etc.
