+++ docs/changelog.rst
@@ -28,7 +28,7 @@ Version 0.3.2
    * User home :code:`~/` is now expanded in paths (`Issue #22 <https://github.com/gpauloski/proxystore/issues/22>`_).
 
 #. Fix :code:`RemoteStore.proxy()` logging the incorrect key (`Issue #26 <https://github.com/gpauloski/proxystore/issues/26>`_).
-#. Failures to resolve a proxy from a :any:`RemoteStore <proxystore.store.remote.RemoteStore>` now raise an :code:`ProxyResolveMissingKey` exception (`Issue #29 <https://github.com/gpauloski/proxystore/issues/29>`_).
+#. Failures to resolve a proxy from a :code:`RemoteStore` now raise an :code:`ProxyResolveMissingKey` exception (`Issue #29 <https://github.com/gpauloski/proxystore/issues/29>`_).
 #. Added prototype stat tracking for stores (`PR #33 <https://github.com/gpauloski/proxystore/pull/33>`_).
 #. CI/Development changes:
 
@@ -67,14 +67,14 @@ Version 0.3.0
 #. :any:`LambdaFactory <proxystore.factory.LambdaFactory>` takes :code:`*args, **kwargs` now. See `Issue #13 <https://github.com/gpauloski/ProxyStore/issues/13>`_. (`#86e6cac <https://github.com/gpauloski/ProxyStore/commit/86e6cac2c782bca7d2ef2e573bd4afc254c4c678>`_)
 #. Added source code button to ReadTheDocs. (`#8a20c20 <https://github.com/gpauloski/ProxyStore/commit/8a20c2099e9eea5235b1dc819ef8c633b21ab662>`_)
 #. Added :any:`FileStore <proxystore.store.file.FileStore>`. (`#e7f8849 <https://github.com/gpauloski/ProxyStore/commit/e7f8849dfd412cb2a451a624ff1fcd001a4615ca>`_)
-#. Added :any:`Store.cleanup() <proxystore.store.base.Store.cleanup>` to store interface for implementations that need to add cleanup logic (such as :any:`FileStore <proxystore.store.file.FileStore>` cleaning up its directory). (`#e7f8849 <https://github.com/gpauloski/ProxyStore/commit/e7f8849dfd412cb2a451a624ff1fcd001a4615ca>`_)
+#. Added :code:`Store.cleanup()` to store interface for implementations that need to add cleanup logic (such as :any:`FileStore <proxystore.store.file.FileStore>` cleaning up its directory). (`#e7f8849 <https://github.com/gpauloski/ProxyStore/commit/e7f8849dfd412cb2a451a624ff1fcd001a4615ca>`_)
 #. Added :any:`GlobusStore <proxystore.store.globus.GlobusStore>` and FuncX+Globus example. (`Issue #15 <https://github.com/gpauloski/ProxyStore/issues/15>`_)
 #. The key is now an optional parameter to :any:`Store.set() <proxystore.store.base.Store.set>` and the official get used by the store will be returned by :any:`Store.set() <proxystore.store.base.Store.set>`. :any:`Store.create_key() <proxystore.store.base.Store.create_key>` can be used to specify key generation when a key is not provided. (`#38a78fa <https://github.com/gpauloski/ProxyStore/commit/38a78fad4ec95012923620523c35e9b9c8083828>`_)
-#. Better inheritance for subclasses of :any:`RemoteStore <proxystore.store.remote.RemoteStore>` and created the base class :any:`RemoteFactory <proxystore.store.remote.RemoteFactory>` to reduce redundant code in other factory types. (`#cf0a631 <https://github.com/gpauloski/ProxyStore/commit/cf0a631646cbec676928daa6a166218185847fa6>`_)
+#. Better inheritance for subclasses of :code:`RemoteStore` and created the base class :code:`RemoteFactory` to reduce redundant code in other factory types. (`#cf0a631 <https://github.com/gpauloski/ProxyStore/commit/cf0a631646cbec676928daa6a166218185847fa6>`_)
 #. Serialization in ProxyStore is now done to bytes rather than strings. (`#d68ac9d <https://github.com/gpauloski/ProxyStore/commit/d68ac9de92cc5d2b902c2fed462e75df7c830c8e>`_)
 #. Added :code:`DEBUG` level logging to :any:`proxystore.store <proxystore.store>`. (`#e0f1052 <https://github.com/gpauloski/ProxyStore/commit/e0f1052a1bae3ccf2af10320852605989b501521>`_)
 #. Better testing infrastructure. Generic store tests now use fixtures to test multiple store types so code is not copy/pasted between tests. Added Globus and Parsl mocking to support unittests for :any:`GlobusStore <proxystore.store.globus.GlobusStore>`. (`#91d3894 <https://github.com/gpauloski/ProxyStore/commit/91d3894bd85de8686fda0d9e425f18e122fa9e82>`_)
-#. Better handling of timestamps in stores that support mutable objects and strict guarantees. Added :any:`RemoteStore.get_timestamp(key) <proxystore.store.remote.RemoteStore.get_timestamp>`. :any:`FileStore <proxystore.store.file.FileStore>` now uses file modified time. (`#e1bbdb8 <https://github.com/gpauloski/ProxyStore/commit/e1bbdb8d485369e86b1a9acef6ccd2c1321c2e8d>`_)
+#. Better handling of timestamps in stores that support mutable objects and strict guarantees. Added :code:`RemoteStore.get_timestamp(key)`. :any:`FileStore <proxystore.store.file.FileStore>` now uses file modified time. (`#e1bbdb8 <https://github.com/gpauloski/ProxyStore/commit/e1bbdb8d485369e86b1a9acef6ccd2c1321c2e8d>`_)
 #. Added the batch methods :any:`Store.set_batch() <proxystore.store.base.Store.set_batch>` and :any:`Store.proxy_batch() <proxystore.store.base.Store.proxy_batch>`. (`#d1d24f7 <https://github.com/gpauloski/ProxyStore/commit/d1d24f76fd8c2e50405d1580f116ac8c7e3d2339>`_)
 #. Various docstring and documentation updates.
