--- docs/source/proxystore.store.remote.rst
@@ -1,7 +1,0 @@
-proxystore.store.remote
-#######################
-
-.. automodule:: proxystore.store.remote
-   :members:
-   :inherited-members:
-   :show-inheritance:
