+++ docs/source/proxystore.store.rst
@@ -9,7 +9,6 @@ proxystore.store
    proxystore.store.globus
    proxystore.store.local
    proxystore.store.redis
-   proxystore.store.remote
    proxystore.store.stats
 
 .. automodule:: proxystore.store
