+++ setup.cfg
@@ -63,9 +63,11 @@ parallel = true
 [flake8]
 docstring-convention = google
 exclude = .git,.github,__pycache__,build,dist,.*egg-info
+ignore-decorator = overrides
 per-file-ignores =
     */__init__.py: F401
     */*_test.py: D10
+    proxystore/store/*.py: D102
     tests/conftest.py: F401
 
 [mypy]
